Added Subscription Controller

+++ Assignment5/src/main/java/com/gonuclei/repository/MasterNewsLetterRepository.java
@@ -3,6 +3,9 @@ package com.gonuclei.repository;
 import com.gonuclei.entities.NewsLetterEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MasterNewsLetterRepository extends CrudRepository<NewsLetterEntity, Integer> {
+/**
+ * The interface Master news letter repository.
+ */
+public interface MasterNewsLetterRepository extends CrudRepository<NewsLetterEntity, Long> {
 
 }
